eclass/java-pkg-simple.eclass: Added code to include *.properties files
Presently java-pkg-simple only compiles *.java files and includes the
resulting *.class files in the jar. Some projects ship *.properties and
other files that need to be included in the jar. This eclass has initial
code to include *.properties files. Which needs to be expanded to a
variable to allow for the package to determine what other types of files
to include in the jar file. This is not done in the best manner,
presently hacking to exclude the parent directory when the file is
copied to the destination, incudling its parents/path. Just not the root
source folder, like src/.